% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinearityR.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{Plot rectangular data as a color-encoded Pearson correlaiton matrix.}
\usage{
corr_heatmap(df)
}
\arguments{
\item{df}{dataframe. 2-D dataframe.}
}
\value{
The heatmap.
}
\description{
The rows and the columns contain variable names, while
the heatmap tiles contain Pearson correlation coefficient and
corresponding colours.
}
\details{
The inputs must be a 2-D data frame
}
\examples{
corr_heatmap(data)
}