% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinearityR.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Calculates Pearson correlation coefficients for all numeric variables.
Rounds the outcome to desired decimals. Outputs a matrix or a dataframe
(longer matrix)}
\usage{
corr_matrix(df, decimals = 2)
}
\arguments{
\item{df}{tibble, the input dataframe}

\item{decimals}{int, the number of decimals}
}
\value{
a generic correlation matrix and its longer form that can be passed to other functions in this package.
}
\description{
Calculates Pearson correlation coefficients for all numeric variables.
Rounds the outcome to desired decimals. Outputs a matrix or a dataframe
(longer matrix)
}
\examples{
corr_matrix(tibble::tibble(x = 1:5, y = 2:6))
}
