% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinearityR.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{Plot rectangular data as a color-encoded Pearson correlaiton matrix.}
\usage{
corr_heatmap(df, color_neg, color_pos)
}
\arguments{
\item{df}{dataframe. 2-D dataframe.}

\item{color_neg}{string. The color for the negative correlation.}

\item{color_pos}{string. The color for the negative correlation.}
}
\value{
The heatmap.
}
\description{
The rows and the columns contain variable names, while
the heatmap tiles contain Pearson correlation coefficient and
corresponding colours.
}
\details{
The inputs must be a 2-D data frame
}
\examples{
corr_heatmap(mtcars[, c(1,3,4,5,6,7)])
}
