% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collinearityR.R
\name{corr_matrix}
\alias{corr_matrix}
\title{Calculate the Pearson correlation coefficients for all numeric variables.
Round the outcome to desired decimals.}
\usage{
corr_matrix(df, decimals = 2)
}
\arguments{
\item{df}{tibble, the input dataframe}

\item{decimals}{int, the number of decimals}
}
\value{
a generic correlation matrix and its longer form
that can be passed to other functions in this package
}
\description{
Calculate the Pearson correlation coefficients for all numeric variables.
Round the outcome to desired decimals.
}
